tutorial/requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
decorator==4.2.1
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
Flask==0.12.2
freeze==1.0.10
gunicorn==19.7.1
httpie==0.9.9
hupper==1.0
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.1
parso==0.1.1
PasteDeploy==1.5.2
pexpect==4.4.0
pickleshare==0.7.4
plaster==1.0
plaster-pastedeploy==0.4.2
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyMySQL==0.8.0
pyramid==1.9.1
pytz==2018.3
range==1.0.0
repoze.lru==0.7
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
translationstring==1.3
uritemplate==3.0.0
urllib3==1.22
venusian==1.1.0
wcwidth==0.1.7
WebOb==1.8.0rc1
Werkzeug==0.14.1
xlrd==1.1.0
zope.deprecation==4.3.0
zope.interface==4.4.3
